Core functionality implemented
All essential functionality to simulate average rate responses over a
range of stimulus and model parameters for an auditory nerve simulation
has been implemented. Most notably, a number of utility functions in
simulation offer a basic set of tools to construct large structured
arrays of simulation parameters that the core of Simulator can pass off
to a multiprocess Pool.