xxx_parameters support for ndarrays
All xxx_parameter functions (where relevant) now provide support for
passing ndarrays instead of lists. This is an excellent quality of life
upgrade for many types of simulations as the structure of numpy arrays
allows for easy post-simulation analysis of results. Tests need to be
added to ensure that these objects can be succesfully passed to run()
and downstream simulation code.