Added ndarray support to sim() and synthesizers
Adjusted some code in sim() and synthesize_sequence() to allow for these
functions to support passing and returning ndarrays... for the most
part, I believe these changes should be sufficient to allow seamless
replacement of lists with ndarrays in most of the existing simulation
code that uses apcmodels. Some additional tests were added to check to
make sure things are working as we expect.